Converting function declarations to arrow functions.
