+++ src/App.css
@@ -3,7 +3,7 @@
 @import './reset.css';
 
 :root {
-  --pink:  #b5179e;
+  --pink: #b5179e;
   --purple: #601293;
   --blue: #3f37c9;
 }
