+++ src/App.js
@@ -3,19 +3,17 @@ import './App.css';
 import Header from './components/Header/Header.js';
 import Footer from './components/Footer/Footer.js';
 import DropDownMenu from './components/DropDownMenu/DropDownMenu';
-import RandomText from './components/RandomText/RandomText';
 
-function App() {
+const App = () => {
   return (
     <div className="app-container">
       <Header />
       <main>
         <DropDownMenu />
-        {/* <RandomText />   */}
       </main>
       <Footer />
     </div>
   );
-}
+};
 
 export default App;
