+++ src/components/Header/Header.module.css
@@ -25,8 +25,6 @@
     #4895ef,
     #4cc9f0
   );
- 
-  
 }
 
 .img {
